devconfig: Wait until cloud instance becomes available
Cloud instance created via terraform may not yet have initialized its
networking stack or may not yet have started the sshd daemon. This causes the
devconfig playbook to fail due to the remote instance not being reachable.

Hence this commit changes ansible to wait for the remote system to be ready
before executing administrative tasks on it.

Signed-off-by: Chandan Babu R <chandan.babu@oracle.com>